refactor(example): root-level glass ripple + isolate board renders
Move RippleProvider out of app/ to components/ripple.tsx. expo-router
treats every file under app/ as a route, so app/ripple.tsx logged
"Route is missing the required default export". The overlay now renders
above the navigator, so the glass wave covers the whole window (native
header included) and works in plain window-space coordinates. It mounts
only while the wave is in flight so it never swallows touches, and the
spring's slow tail is torn down via a progress reaction (no frozen frame).

Wrap <Chessboard> in React.memo so the chrome's per-move state updates
(status / moves / captured) no longer reconcile the board subtree.
